+++ Sources/TuistCache/ContentHashing/TargetScriptsContentHasher.swift
@@ -4,7 +4,7 @@ import TuistCore
 import TuistGraph
 
 public protocol TargetScriptsContentHashing {
-    func hash(targetScripts: [TargetScript]) throws -> String
+    func hash(targetScripts: [TargetScript], sourceRootPath: AbsolutePath) throws -> String
 }
 
 /// `TargetScriptsContentHasher`
@@ -22,33 +22,30 @@ public final class TargetScriptsContentHasher: TargetScriptsContentHashing {
 
     /// Returns the hash that uniquely identifies an array of target scripts
     /// The hash takes into consideration the content of the script to execute, the content of input/output files, the name of the tool to execute, the order, the arguments and its name
-    public func hash(targetScripts: [TargetScript]) throws -> String {
+    public func hash(targetScripts: [TargetScript], sourceRootPath: AbsolutePath) throws -> String {
         var stringsToHash: [String] = []
         for script in targetScripts {
             var pathsToHash: [AbsolutePath] = []
             script.path.map { pathsToHash.append($0) }
-            let scriptPaths = script.inputPaths + script.inputFileListPaths + script.outputPaths + script.outputFileListPaths
-            scriptPaths.forEach { path in
+            (script.inputPaths + script.inputFileListPaths).forEach { path in
                 if path.pathString.contains("$") {
-                    stringsToHash.append(path.pathString)
-                    logger
-                        .notice(
-                            "The path of the file \'\(path.url.lastPathComponent)\' is hashed, not the content. Because it has a build variable."
-                        )
+                    stringsToHash.append(path.relative(to: sourceRootPath).pathString)
+                    logger.notice(
+                        "The path of the file \'\(path.url.lastPathComponent)\' is hashed, not the content. Because it has a build variable."
+                    )
                 } else {
                     pathsToHash.append(path)
                 }
             }
-            let fileHashes = try pathsToHash.map { try contentHasher.hash(path: $0) }
+            stringsToHash.append(contentsOf: try pathsToHash.map { try contentHasher.hash(path: $0) })
             stringsToHash.append(
-                contentsOf: fileHashes +
-                    [
-                        script.name,
-                        script.tool ?? "", // TODO: don't default to ""
-                        script.order.rawValue,
-                    ] +
-                    script.arguments
+                contentsOf: (script.outputPaths + script.outputFileListPaths).map { $0.relative(to: sourceRootPath).pathString }
             )
+            stringsToHash.append(contentsOf: [
+                script.name,
+                script.tool ?? "",
+                script.order.rawValue,
+            ] + script.arguments)
         }
         return try contentHasher.hash(stringsToHash)
     }