+++ Tests/TuistGeneratorTests/Generator/ConfigGeneratorTests.swift
@@ -19,7 +19,6 @@ final class ConfigGeneratorTests: TuistUnitTestCase {
         super.setUp()
         pbxproj = PBXProj()
         pbxTarget = PBXNativeTarget(name: "Test")
-        system.swiftVersionStub = { "5.2" }
         pbxproj.add(object: pbxTarget)
         subject = ConfigGenerator()
     }
@@ -67,7 +66,7 @@ final class ConfigGeneratorTests: TuistUnitTestCase {
             "INFOPLIST_FILE": "$(SRCROOT)/Info.plist",
             "PRODUCT_BUNDLE_IDENTIFIER": "com.test.bundle_id",
             "CODE_SIGN_ENTITLEMENTS": "$(SRCROOT)/Test.entitlements",
-            "SWIFT_VERSION": "5.2",
+            "SWIFT_VERSION": "5.0",
         ]
 
         let debugSettings = [