+++ projects/cloud/app/graphql/types/owner_type.rb
@@ -4,11 +4,13 @@ module Types
   class OwnerType < Types::BaseUnion
     possible_types UserType, OrganizationType
 
-    def self.resolve_type(object, context)
-      if object.is_a?(User)
-        Types::UserType
-      elsif object.is_a?(Organization)
-        Types::OrganizationType
+    class << self
+      def resolve_type(object, context)
+        if object.is_a?(User)
+          Types::UserType
+        elsif object.is_a?(Organization)
+          Types::OrganizationType
+        end
       end
     end
   end