+++ projects/cloud/app/lib/environment.rb
@@ -3,13 +3,15 @@
 module Environment
   TRUTHY_VALUES = ["1", "true", "TRUE", "yes", "YES"]
 
-  def self.use_env_variables?(env: ENV)
-    truthy?(env["TUIST_CLOUD_ENV_VARIABLES"])
-  end
+  class << self
+    def use_env_variables?(env: ENV)
+      truthy?(env["TUIST_CLOUD_ENV_VARIABLES"])
+    end
 
-  def self.truthy?(value)
-    return false if value.blank?
+    def truthy?(value)
+      return false if value.blank?
 
-    TRUTHY_VALUES.any? { |v| v == value.to_s }
+      TRUTHY_VALUES.any? { |v| v == value.to_s }
+    end
   end
 end